Pull up reusable extensions to `gradle-api` module
