+++ gradle-api/build.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
 
     implementation(ToolBase.pluginBase)
     implementation(project(":api"))
+    implementation(project(":params"))
 
     testImplementation(gradleApi())
     testImplementation(ToolBase.pluginTestlib)
